Started Warning Players
WARNING: DO NOT PULL FROM THIS. This edited some code not to work quite in sync with the rest of the game.

This is preliminary testing in order to make sure everything is set up correctly. Spawners now change color and the enemies will later spawn according to color scheme.