Match3 kinda works?
How I did it:
There is a minimum match number. If the matching logic detects that a match over 3 was made, then it matches and then respawns the matches at the top.
Gravity does the rest.

Problems/Bugs:
Right now, it can't distinguish rows and columns so it'll try and match up/down and left/right together.

It also uses 3d and gravity physics in order to work. We can work around this and it's more of a minor gripe if anything.

Does not return type so that Zach can increment to his defense/offense.